+++ .flutter-plugins-dependencies
@@ -1,0 +1,1 @@
+{"_info":"// This is a generated file; do not edit or check into version control.","dependencyGraph":[{"name":"cloud_firestore","dependencies":["firebase_core"]},{"name":"firebase_analytics","dependencies":[]},{"name":"firebase_core","dependencies":[]},{"name":"fluttertoast","dependencies":[]}]}
