+++ lib/main.dart
@@ -1,527 +1,5 @@
-import 'dart:async';
-import 'package:cloud_firestore/cloud_firestore.dart';
-import 'package:flutter/cupertino.dart';
-import 'package:flutter/material.dart';
-import 'package:fluttertoast/fluttertoast.dart';
-import 'package:flutter/services.dart';
+import 'package:agrimaster_app/Routes.dart';
 
 void main() {
-  runApp(new MaterialApp(
-    theme: ThemeData(
-        buttonTheme: ButtonThemeData(minWidth: 12),
-        primarySwatch: Colors.green
-    ),
-    debugShowCheckedModeBanner: false,  //デバッグリボン非表示
-    title: 'Navigation with Routes',
-    routes: <String, WidgetBuilder>{
-      '/': (_) => new Splash(),
-      '/login': (_) => new Login(),
-      '/registration': (_) => new Registration(),
-      '/home': (_) => new Home(),
-      '/setting': (_) => new Setting(),
-    },
-  ));
-}
-
-// --------------
-// スプラッシュ画面
-// --------------
-class Splash extends StatefulWidget {
-  @override
-  _SplashState createState() => new _SplashState();
-}
-
-class _SplashState extends State<Splash> {
-  @override
-  void initState() {
-    super.initState();
-
-    new Future.delayed(const Duration(seconds: 3))
-        .then((value) => handleTimeout());
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return new Scaffold(
-      body: new Center(
-        // TODO: スプラッシュアニメーション
-        child: const CircularProgressIndicator(),
-      ),
-    );
-  }
-
-  void handleTimeout() {
-    // ログイン画面へ
-    Navigator.of(context).pushReplacementNamed("/login");
-  }
-}
-
-// -----------
-// ログイン画面
-// -----------
-class Login extends StatefulWidget {
-  @override
-  _LoginState createState() => new _LoginState();
-}
-
-class _LoginState extends State<Login> {
-  @override
-  Widget build(BuildContext context) {
-    return new Scaffold(
-      appBar: new AppBar(
-        title: const Text("AGRIMASTERにログイン"),
-      ),
-      body: new Center(
-        child: new Form(
-          child: new SingleChildScrollView(
-            padding: const EdgeInsets.symmetric(horizontal: 16.0),
-            child: new Column(
-              crossAxisAlignment: CrossAxisAlignment.stretch,
-              children: <Widget>[
-                const SizedBox(height: 24.0),
-                new TextFormField(
-                  maxLength: 30,
-                  decoration: const InputDecoration(
-                    border: const UnderlineInputBorder(),
-                    labelText: 'メールアドレス',
-                  ),
-                ),
-                const SizedBox(height: 24.0),
-                new TextFormField(
-                  maxLength: 16,
-                  decoration: new InputDecoration(
-                    border: const UnderlineInputBorder(),
-                    labelText: 'パスワード',
-                  ),
-                ),
-                const SizedBox(height: 24.0),
-                new Center(
-                  child: new RaisedButton(
-                    child: const Text('ログイン'),
-                    onPressed: () {
-                      // TODO: ログイン処理
-                      // ホーム画面へ
-                      Navigator.of(context).pushReplacementNamed("/home");
-                    },
-                  ),
-                ),
-                const SizedBox(height: 24.0),
-                Center(
-                  child: new RaisedButton(
-                  child: const Text("新規登録"),
-                  onPressed: () {
-                    // 登録画面へ
-                    Navigator.of(context).pushNamed("/registration");
-                  },
-                ),
-                )
-              ],
-            ),
-          ),
-        ),
-      ),
-    );
-  }
-}
-
-// ---------
-// 登録画面
-// ---------
-
-class Registration extends StatefulWidget {
-  @override
-  _RegistrationState createState() => new _RegistrationState();
-}
-
-class _RegistrationState extends State<Registration> {
-  @override
-  Widget build(BuildContext context) {
-    return new Scaffold(
-      appBar: new AppBar(
-        title: const Text("新規登録"),
-      ),
-      body: new Center(
-        child: new Form(
-          child: new SingleChildScrollView(
-            padding: const EdgeInsets.symmetric(horizontal: 16.0),
-            child: new Column(
-              crossAxisAlignment: CrossAxisAlignment.stretch,
-              children: <Widget>[
-                const SizedBox(height: 24.0),
-                new TextFormField(
-                  maxLength: 30,
-                  decoration: const InputDecoration(
-                    border: const UnderlineInputBorder(),
-                    labelText: 'ユーザ名',
-                  ),
-                ),
-                const SizedBox(height: 24.0),
-                new TextFormField(
-                  maxLength: 30,
-                  decoration: const InputDecoration(
-                    border: const UnderlineInputBorder(),
-                    labelText: 'メールアドレス',
-                  ),
-                ),
-                const SizedBox(height: 24.0),
-                new TextFormField(
-                  maxLength: 16,
-                  decoration: new InputDecoration(
-                    border: const UnderlineInputBorder(),
-                    labelText: 'パスワード',
-                  ),
-                ),
-                const SizedBox(height: 24.0),
-                new Center(
-                  child: new RaisedButton(
-                    child: const Text('新規登録'),
-                    onPressed: () {
-                      // TODO: 新規登録処理
-
-                      // ホーム画面へ
-                      Navigator.of(context).pushReplacementNamed("/login");
-                    },
-                  ),
-                ),
-                const SizedBox(height: 24.0),
-                Center(
-                  child: new RaisedButton(
-                  child: const Text("戻る"),
-                  onPressed: () {
-                    // ログイン画面へ戻る
-                    Navigator.of(context).pop();
-                    //Navigator.of(context).pushNamed("/login");
-                  },
-                ),
-                )
-              ],
-            ),
-          ),
-        ),
-      ),
-    );
-  }
-}
-
-// ---------
-// ホーム画面
-// ---------
-class Home extends StatelessWidget {
-  @override
-  Widget build(BuildContext context) {
-    return new Scaffold(
-      appBar: new AppBar(
-        title: const Text("ホーム"),
-        actions: <Widget>[      // Add 3 lines from here...
-          IconButton(icon: Icon(Icons.settings), onPressed:(){
-            Navigator.of(context).pushNamed("/setting");
-          } ),
-        ], 
-      ),
-      body: new Center(
-        child: new Column(
-          children: <Widget>[
-
-            Padding(
-              padding: EdgeInsets.only(top: 70.0, bottom: 10.0),
-              child: const Text('温度',
-              style: TextStyle(fontSize: 20)),
-            ),
-
-            Padding(
-              padding: EdgeInsets.only(top: 10.0, bottom: 30.0),
-              child: const Text('20℃',
-              style: TextStyle(fontSize: 40)),
-            ),
-
-            Padding(
-              padding: EdgeInsets.only(top: 20.0, bottom: 10.0),
-              child: const Text('湿度',
-              style: TextStyle(fontSize: 20)),
-            ),
-
-            Padding(
-              padding: EdgeInsets.only(top: 10.0, bottom: 40.0),
-              child: const Text('50%',
-              style: TextStyle(fontSize: 40)),
-            ),
-
-            new RaisedButton(
-            child: const Text("設定"),
-            onPressed: () {
-              // 設定画面へ
-              Navigator.of(context).pushNamed("/setting");
-              },
-            ),
-          ]
-        ),
-      ),
-    );
-  }
-}
-
-// ---------
-// 設定画面
-// ---------
-class Setting extends StatefulWidget {
-  @override
-  _SettingState createState() => new _SettingState();
-}
-
-class DisableKeybord extends FocusNode {
-  @override
-  bool get hasFocus => false;
-}
-
-class _SettingState extends State<Setting> {
-
-  bool _active = false;
-  void _changeSwitch(bool e) => setState(() => _active = e);
-  var _start = '';
-  var _end = '';
-  var _rangeValues = RangeValues(40.0, 60.0);
-
-  _updateLabels(RangeValues values) {
-    _start = '${_rangeValues.start.round()}';
-    _end = '${_rangeValues.end.round()}';
-  }
-
-  @override
-  void initState() {
-    _updateLabels(_rangeValues);
-
-    isPasswordVisible = false;
-    isConfirmPasswordVisible = false;
-    super.initState();
-  }
-
-  bool isPasswordVisible = false;
-  bool isConfirmPasswordVisible = false;
-
-  GlobalKey<ScaffoldState> screen = GlobalKey<ScaffoldState>();
-  
-  /*
-  final GlobalKey<ScaffoldState> _scaffoldstate = new GlobalKey<ScaffoldState>();
-
-  void _showBar(){
-    _scaffoldstate.currentState.showSnackBar(new SnackBar(content: new Text('設定温度を確定しました！')));
-  }
-
-   */
-
-  @override
-  Widget build(BuildContext context) {
-    return new Scaffold(
-      key: screen,
-      //key: _scaffoldstate,
-        appBar: new AppBar(
-            title: const Text("設定"),
-        ),
-        body: new Center(
-            child: new Column(
-              children: <Widget>[
-
-                Padding(
-                  padding: EdgeInsets.only(top: 20, bottom: 10, left: 5),
-                    child: MergeSemantics(
-                      child: ListTile(
-                        title: Text('通知',
-                                style: TextStyle(fontSize:  23)),
-                        trailing: CupertinoSwitch(
-                          value: _active,
-                          onChanged: _changeSwitch,
-                        ),
-                      ),
-                    ),
-                ),
-
-                Align(
-                  alignment: Alignment.centerLeft,
-                  child: Padding(
-                    padding: EdgeInsets.only(bottom: 35.0, left: 20),
-                    child: Text("温度を設定", style: TextStyle(fontSize: 23)),
-                  ),
-                ),
-                SliderTheme(
-                    data: SliderThemeData(
-                    activeTrackColor: Colors.green[250],
-                    inactiveTrackColor: Colors.grey,
-                    showValueIndicator: ShowValueIndicator.always,
-                    ),
-                    child: RangeSlider(
-                        labels: RangeLabels(_start, _end),
-                        values: _rangeValues,
-                        min: 1,
-                        max: 100,
-                        divisions: 100,
-                        onChanged: (values) {
-                          _rangeValues = values;
-                          setState(() => _updateLabels(values));
-                        },
-                    ),
-                ),
-
-                Container(
-                    padding: EdgeInsets.only(bottom: 5),
-                      child: Row(
-                          mainAxisAlignment: MainAxisAlignment.spaceAround,
-                          children: <Widget>[
-                            Column(
-                              children: <Widget>[
-                                Text(
-                                  '$_start℃',
-                                    style: TextStyle(fontSize: 40),
-                                ),
-                                Text('下限'),
-                              ],
-                            ),
-                            Column(
-                              children: <Widget>[
-                                Text(
-                                  '$_end℃',
-                                    style: TextStyle(fontSize: 40),
-                                ),
-                                Text('上限'),
-                              ],
-                            ),
-                          ],
-                      ),
-                ),
-
-                RaisedButton(
-                  //color: Colors.blueAccent,
-                  shape: StadiumBorder(),
-                  onPressed: () {
-                    //TODO:設定温度送信
-                    screen.currentState.removeCurrentSnackBar();
-                    screen.currentState
-                        .showSnackBar(SnackBar(
-                          content: Text('温度設定を確定しました！')));
-                  },
-                  child: Text('温度確定',
-                            style: TextStyle(color: Colors.white)
-                            ),
-                ),
-
-                Container(
-                  alignment: Alignment.centerLeft,
-                    padding: EdgeInsets.only(top: 20, left: 20),
-                    child: Text("アカウント", 
-                            style: TextStyle(fontSize: 25)),
-                ),
-
-                Container(
-                  padding: EdgeInsets.only(top: 10, bottom: 10, left: 20, right: 20),
-                  child: TextFormField(
-                    initialValue: 'test@mail.com',
-                    focusNode: DisableKeybord(),
-                    maxLines: 1,
-                    autofocus: false,
-                    decoration: new InputDecoration(
-                      hintText: 'メールアドレス',
-                      icon: new Icon(
-                        Icons.mail,
-                        color: Colors.grey,
-                      )
-                    ),
-                  ),
-                ),
-
-                Container(
-                  padding: EdgeInsets.only(top: 10, bottom: 10, left: 20, right: 20),
-                  child: TextFormField(
-                    initialValue: 'testPasswordDAYO!',
-                    focusNode: DisableKeybord(),
-                    maxLines: 1,
-                    obscureText: !isConfirmPasswordVisible,
-                    autofocus: false,
-                    decoration: new InputDecoration(
-                      hintText: 'Password',
-                      suffixIcon: IconButton(
-                          icon: Icon(isConfirmPasswordVisible
-                              ? Icons.visibility
-                              : Icons.visibility_off),
-                      onPressed: () {
-                        setState(() {
-                          isConfirmPasswordVisible =
-                                  !isConfirmPasswordVisible;
-                        });
-                      },
-                    ),
-                    icon: Icon(Icons.lock)
-                    ),
-                  ),
-                ),
-
-
-                //const SizedBox(height: 24.0),
-                /*new RaisedButton(
-                    child: const Text("ホーム"),
-                        onPressed: () {
-                // ホーム画面へ戻る　
-                            Navigator.popUntil(context, ModalRoute.withName("/home"));
-                        },
-                ),
-
-                 */
-                const SizedBox(height: 20.0),
-                new RaisedButton(
-                      color: Colors.redAccent,
-                      shape: StadiumBorder(),
-                      child: const Text("ログアウト",
-                                    style: TextStyle(color: Colors.white)),
-                        onPressed: () {
-                    // 確認ダイアログ表示
-                            showDialog<bool>(
-                                context: context,
-                                builder: (BuildContext context) {
-                                  return new AlertDialog(
-                                    shape: RoundedRectangleBorder(
-                                      borderRadius: BorderRadius.circular(20)
-                                    ),
-                                    content: const Text('ログアウトしてもよろしいですか？'),
-                                    actions: <Widget>[
-                                      new RaisedButton(
-                                            color: Colors.red,
-                                            shape: StadiumBorder(),
-                                            child: const Text('いいえ', 
-                                                          style: TextStyle(color: Colors.white),),
-                                            onPressed: () {
-                                              // 引数をfalseでダイアログ閉じる
-                                              Navigator.of(context).pop(false);
-                                            },
-                                        ),
-                                        new RaisedButton(
-                                              color: Colors.blueAccent,
-                                              shape: StadiumBorder(),
-                                              child: const Text('はい', 
-                                                            style: TextStyle(color: Colors.white),),
-                                              onPressed: () {
-                                                //TODO:ログアウト
-                                                // 引数をtrueでダイアログ閉じる
-                                                Navigator.of(context).pop(true);
-                                              },
-                                          ),
-                                      ],
-                                    );
-                                  },
-                                  ).then<void>((aBool) {
-                                    // ダイアログがYESで閉じられたら...
-                                    if (aBool) {
-                                      // 画面をすべて除いてスプラッシュを表示
-                                      Navigator.pushAndRemoveUntil(
-                                        context,
-                                        new MaterialPageRoute(
-                                              builder: (context) => new Splash()),
-                                              (_) => false
-                                      );
-                                    }
-                                  });
-                                },
-                ),
-          ],
-        ),
-      ),
-    );
-  }
+  new Routes();
 }
