+++ lib/screens/Login/index.dart
@@ -1,0 +1,72 @@
+import 'dart:async';
+import 'package:cloud_firestore/cloud_firestore.dart';
+import 'package:flutter/cupertino.dart';
+import 'package:flutter/material.dart';
+import 'package:fluttertoast/fluttertoast.dart';
+import 'package:flutter/services.dart';
+
+
+class Login extends StatefulWidget {
+  @override
+  _LoginState createState() => new _LoginState();
+}
+
+class _LoginState extends State<Login> {
+  @override
+  Widget build(BuildContext context) {
+    return new Scaffold(
+      appBar: new AppBar(
+        title: const Text("AGRIMASTERにログイン"),
+      ),
+      body: new Center(
+        child: new Form(
+          child: new SingleChildScrollView(
+            padding: const EdgeInsets.symmetric(horizontal: 16.0),
+            child: new Column(
+              crossAxisAlignment: CrossAxisAlignment.stretch,
+              children: <Widget>[
+                const SizedBox(height: 24.0),
+                new TextFormField(
+                  maxLength: 30,
+                  decoration: const InputDecoration(
+                    border: const UnderlineInputBorder(),
+                    labelText: 'メールアドレス',
+                  ),
+                ),
+                const SizedBox(height: 24.0),
+                new TextFormField(
+                  maxLength: 16,
+                  decoration: new InputDecoration(
+                    border: const UnderlineInputBorder(),
+                    labelText: 'パスワード',
+                  ),
+                ),
+                const SizedBox(height: 24.0),
+                new Center(
+                  child: new RaisedButton(
+                    child: const Text('ログイン'),
+                    onPressed: () {
+                      // TODO: ログイン処理
+                      // ホーム画面へ
+                      Navigator.of(context).pushReplacementNamed("/home");
+                    },
+                  ),
+                ),
+                const SizedBox(height: 24.0),
+                Center(
+                  child: new RaisedButton(
+                  child: const Text("新規登録"),
+                  onPressed: () {
+                    // 登録画面へ
+                    Navigator.of(context).pushNamed("/registration");
+                  },
+                ),
+                )
+              ],
+            ),
+          ),
+        ),
+      ),
+    );
+  }
+}
