+++ lib/screens/Splash/index.dart
@@ -1,0 +1,34 @@
+import 'dart:async';
+import 'package:flutter/cupertino.dart';
+import 'package:flutter/material.dart';
+
+
+class Splash extends StatefulWidget {
+  @override
+  _SplashState createState() => new _SplashState();
+}
+
+class _SplashState extends State<Splash> {
+  @override
+  void initState() {
+    super.initState();
+
+    new Future.delayed(const Duration(seconds: 3))
+        .then((value) => handleTimeout());
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return new Scaffold(
+      body: new Center(
+        // TODO: スプラッシュアニメーション
+        child: const CircularProgressIndicator(),
+      ),
+    );
+  }
+
+  void handleTimeout() {
+    // ログイン画面へ
+    Navigator.of(context).pushReplacementNamed("/login");
+  }
+}
